+++ src/utils/http.nim
@@ -8,6 +8,7 @@
 import std/[
   httpclient,
   httpcore,
+  strutils,
 ]
 
 export httpclient
@@ -17,3 +18,15 @@ proc update*(self: HttpHeaders, with: HttpHeaders): HttpHeaders =
   for k, v in with.pairs():
     self[k] = v
   return self
+
+proc mustGet*(headers: HttpHeaders, header: string, msg: string): string =
+  if not headers.hasKey(header):
+    raise newException(ValueError, msg)
+  return headers[header]
+
+proc mustGetInt*(headers: HttpHeaders, header: string, msg: string): int =
+  let value = headers.mustGet(header, msg)
+  try:
+    return parseInt(value)
+  except:
+    raise newException(ValueError, msg & ": invalid integer: " & value)